Configured Webpack config using typescript
