+++ .webpack/common/entry.ts
@@ -1,0 +1,5 @@
+import { APP_INIT_FILE } from '../constants';
+
+export default () => ({
+  main: APP_INIT_FILE,
+})
