+++ .webpack/common/index.ts
@@ -1,0 +1,7 @@
+export {default as devServer} from './devServer';
+export {default as entry} from './entry';
+export {default as module} from './module';
+export {default as optimization} from './optimization';
+export {default as output} from './output';
+export {default as plugins} from './plugins';
+export {default as resolve} from './resolve';
