+++ .webpack/common/module/index.ts
@@ -1,0 +1,7 @@
+import { ModuleOptions } from 'webpack';
+import rules from './rules';
+
+export default (): ModuleOptions => ({
+  rules: rules(),
+  strictExportPresence: true,
+})
