+++ .webpack/types.ts
@@ -1,0 +1,21 @@
+export type EnvConfig = {
+  DOMAIN: string,
+  PROXIES: Array<string>,
+  TOKEN: string,
+}
+
+export enum EnvType {
+  dev = 'dev',
+  stage = 'stage',
+  prod = 'prod',
+}
+
+export type WebpackConfig = {
+  WEBPACK_SERVE?: boolean;
+  env?: EnvType;
+  url?: string;
+  isDev?: boolean;
+  envConfig?: EnvConfig;
+}
+
+export type ConfigBuilder = (config: WebpackConfig) => any;
