--- src/domain/index.ts
@@ -1,3 +1,0 @@
-export * as Auth from './Auth';
-export * as User from './User';
-export * as Settings from './Settings';
