+++ src/ui/hooks/index.ts
@@ -1,7 +1,5 @@
 import {useEffect} from 'react';
 
-export * from './di';
-
 export function useDidMount(callback: (() => () => void) | (() => void)): void {
   useEffect(() => {
     return callback();
