--- src/ui/modules/Settings/index.ts
@@ -1,1 +1,0 @@
-export {Theme} from './Theme';
