--- src/ui/modules/index.ts
@@ -1,1 +1,0 @@
-export * as Settings from './Settings';
