--- src/ui/pages/Auth/SignIn/types.ts
@@ -1,5 +1,0 @@
-export const APPLICATION_NAME = 'Product Inventory Engine'
-
-export const EMAIL_LABEL = 'Email'
-export const PASSWORD_LABEL = 'Password'
-export const SIGN_IN_LABEL = 'Sing In'
