--- src/ui/pages/Auth/index.tsx
@@ -1,1 +1,0 @@
-export * from './SignIn'
