--- src/ui/pages/index.ts
@@ -1,2 +1,0 @@
-export * as Auth from './Auth';
-export * from './Home';
